fix(editor): Move smart decimals to number utils
