+++ packages/@n8n/utils/src/number/smartDecimal.test.ts
