+++ packages/@n8n/utils/src/number/smartDecimal.ts
